refactor(security): move user classes and interfaces
