--- security-service/src/main/java/com/ilogos/security/user/UserProjection.java
@@ -1,9 +1,0 @@
-package com.ilogos.security.user;
-
-import java.util.Set;
-
-import com.ilogos.security.user.common.UserView;
-
-public interface UserProjection extends UserView {
-    Set<String> getRoles();
-}
