+++ security-service/src/main/java/com/ilogos/security/user/UserService.java
@@ -14,9 +14,9 @@ import org.springframework.stereotype.Service;
 import com.ilogos.security.auth.AuthService;
 import com.ilogos.security.exception.ExceptionWithStatus;
 import com.ilogos.security.user.UserController.UpdateUserRequest;
-import com.ilogos.security.user.common.RoleType;
 import com.ilogos.security.user.emailHistory.EmailHistory;
 import com.ilogos.security.user.emailHistory.EmailHistoryRepository;
+import com.ilogos.security.user.model.RoleType;
 import com.ilogos.security.user.usernameHistory.UsernameHistory;
 import com.ilogos.security.user.usernameHistory.UsernameHistoryRepository;
 import com.ilogos.security.utils.TokenInfo;
