--- security-service/src/main/java/com/ilogos/security/user/common/UserView.java
@@ -1,7 +1,0 @@
-package com.ilogos.security.user.common;
-
-import java.time.Instant;
-
-public interface UserView extends UserMinimalView {
-    Instant getLastTokenIssuedAt();
-}
