+++ security-service/src/main/java/com/ilogos/security/user/model/IUser.java
@@ -1,0 +1,7 @@
+package com.ilogos.security.user.model;
+
+import java.time.Instant;
+
+public interface IUser extends IUserBase {
+    Instant getLastTokenIssuedAt();
+}
