+++ security-service/src/main/java/com/ilogos/security/user/model/RoleType.java
@@ -1,4 +1,4 @@
-package com.ilogos.security.user.common;
+package com.ilogos.security.user.model;
 
 import org.springframework.security.core.GrantedAuthority;
