+++ security-service/src/main/java/com/ilogos/security/user/model/UserDTO.java
@@ -1,17 +1,17 @@
-package com.ilogos.security.user;
+package com.ilogos.security.user.model;
 
 import java.time.Instant;
 import java.util.Set;
 import java.util.UUID;
 
-import com.ilogos.security.user.common.UserView;
+import com.ilogos.security.user.User;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class UserDTO implements UserView {
+public class UserDTO implements IUser {
 
     private final UUID id;
     private final String username;
@@ -19,7 +19,7 @@ public class UserDTO implements UserView {
     private final Set<String> roles;
     private final Instant lastTokenIssuedAt;
 
-    private static UserDTO from(UserView view, Set<String> roles) {
+    private static UserDTO from(IUser view, Set<String> roles) {
         return new UserDTO(
                 view.getId(),
                 view.getUsername(),
