+++ security-service/src/main/java/com/ilogos/security/user/model/UserProjection.java
@@ -1,0 +1,7 @@
+package com.ilogos.security.user.model;
+
+import java.util.Set;
+
+public interface UserProjection extends IUser {
+    Set<String> getRoles();
+}
